requirements.txt
streamlit
tensorflow
pillow
numpy
pandas
scikit-learn
transformers
torch
spotipy
matplotlib
seaborn
